feat: embed core, refactor installer, add comprehensive tests
- Embed betterauth-core directly as core-embedded/ (no separate dependency)
- Refactor InstallCommand into modular Installer/ components
- Add OAuth integration layer (src/OAuth/)
- Add DI compiler passes
- Add UserAdapter for Security integration
- Add comprehensive test suite (Command, Controller, Core, Functional, Installer, Storage)
- Update README with unified ecosystem documentation
- Sync core-embedded script for Laravel parity

+++ src/License/LicenseValidator.php
@@ -152,7 +152,7 @@ final class LicenseValidator
         $duration = $gracePeriodData['duration'] ?? self::DEFAULT_GRACE_PERIOD;
         $elapsed = time() - $startedAt;
 
-        return max(0, $duration - $elapsed);
+        return (int) max(0, $duration - $elapsed);
     }
 
     /**
